requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
streamlit==1.22.0
torch==2.0.0
torchvision==0.15.1
transformers==4.28.1
Pillow==9.5.0
numpy<2
matplotlib==3.7.1
pandas==2.0.0
requests==2.28.2
opencv-python-headless==4.7.0.72
scikit-image
python-multipart